# Conflicts:
#	common-repository/src/main/java/cn/qkl/common/repository/Tables.java
#	common-repository/src/main/resources/generatorConfig.xml
#	webserver/src/main/java/cn/qkl/webserver/controller/NotificationController.java
#	webserver/src/main/java/cn/qkl/webserver/dao/ThresholdsDao.java
#	webserver/src/main/java/cn/qkl/webserver/dto/noticification/ReceiveTimeDTO.java
#	webserver/src/main/java/cn/qkl/webserver/service/NotificationService.java

+++ common-repository/src/main/java/cn/qkl/common/repository/mapper/EvidenceWebDynamicSqlSupport.java
@@ -56,10 +56,10 @@ public final class EvidenceWebDynamicSqlSupport {
     public static final SqlColumn<String> packOssPath = evidenceWeb.packOssPath;
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    public static final SqlColumn<Long> evidencePhase = evidenceWeb.evidencePhase;
+    public static final SqlColumn<Integer> evidencePhase = evidenceWeb.evidencePhase;
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
-    public static final SqlColumn<Long> evidenceType = evidenceWeb.evidenceType;
+    public static final SqlColumn<Integer> evidenceType = evidenceWeb.evidenceType;
 
     @Generated("org.mybatis.generator.api.MyBatisGenerator")
     public static final SqlColumn<String> hash = evidenceWeb.hash;
@@ -111,9 +111,9 @@ public final class EvidenceWebDynamicSqlSupport {
 
         public final SqlColumn<String> packOssPath = column("pack_oss_path", JDBCType.VARCHAR);
 
-        public final SqlColumn<Long> evidencePhase = column("evidence_phase", JDBCType.BIGINT);
+        public final SqlColumn<Integer> evidencePhase = column("evidence_phase", JDBCType.INTEGER);
 
-        public final SqlColumn<Long> evidenceType = column("evidence_type", JDBCType.BIGINT);
+        public final SqlColumn<Integer> evidenceType = column("evidence_type", JDBCType.INTEGER);
 
         public final SqlColumn<String> hash = column("hash", JDBCType.VARCHAR);
